add test for SendSurveyNotificationJob and namespace it under StateFile
