+++ app/jobs/state_file/send_survey_notification_job.rb
@@ -1,0 +1,16 @@
+module StateFile
+  class SendSurveyNotificationJob < ApplicationJob
+    def perform(intake, submission)
+      StateFile::MessagingService.new(
+        intake: intake,
+        submission: submission,
+        message: StateFile::AutomatedMessage::SurveyNotification,
+        body_args: { survey_link: StateFile::StateInformationService.survey_link(intake.state_code) }
+      ).send_message
+    end
+
+    def priority
+      PRIORITY_LOW
+    end
+  end
+end
