+++ spec/services/state_file/after_transition_messaging_service_spec.rb
@@ -77,11 +77,11 @@ describe StateFile::AfterTransitionMessagingService do
   end
 
   describe '#schedule_survey_notification_job' do
-    it 'enqueues SendSurveyNotificationJob' do
+    it 'enqueues StateFile::SendSurveyNotificationJob' do
       frozen_time = Time.now
       Timecop.freeze(frozen_time) do
-        expect(SendSurveyNotificationJob).to receive(:set).with(wait_until: (frozen_time + 23.hours)).and_return(SendSurveyNotificationJob)
-        expect(SendSurveyNotificationJob).to receive(:perform_later).with(intake, efile_submission)
+        expect(StateFile::SendSurveyNotificationJob).to receive(:set).with(wait_until: (frozen_time + 23.hours)).and_return(StateFile::SendSurveyNotificationJob)
+        expect(StateFile::SendSurveyNotificationJob).to receive(:perform_later).with(intake, efile_submission)
 
         messaging_service.schedule_survey_notification_job
       end
