+++ spec/services/state_file/reminder_to_finish_state_return_service_spec.rb
@@ -1,7 +1,6 @@
 require 'rails_helper'
 
 describe StateFile::ReminderToFinishStateReturnService do
-
   describe ".run" do
     let(:message) { StateFile::AutomatedMessage::FinishReturn }
     let(:state_file_messaging_service) { StateFile::MessagingService.new(intake: intake, message: message) }
